Reorganized the code, cleaned up some useless bits, and adding documentation comments

+++ include/tasks/Exam.hpp
@@ -1,18 +1,23 @@
 #ifndef EXAM_H
 #define EXAM_H
 
+// Task header file
 #include "tasks/GradedTask.hpp"
+
+// Built in libraries
 #include <string>
 #include <cstdint>
 
+
 class Exam : public GradedTask {
     public:
+        // Constructor
         Exam(
             int idNum, 
             const std::string& name, 
             const std::string& date, 
             bool status,
-            uint8_t grade
+            uint16_t grade
         );
 };
