+++ include/tasks/GradedTask.hpp
@@ -1,24 +1,31 @@
 #ifndef GRADEDTASK_H
 #define GRADEDTASK_H
 
+// Task header file
 #include "tasks/Task.hpp"
+
+// Built in libraries
 #include <string>
 #include <cstdint>
 
+
 class GradedTask : public Task {
     protected:
-        uint8_t grade;
+        // Field
+        uint16_t grade;
     public:
+        // Constructor & Deconstructor
         GradedTask(
             int idNum, 
             const std::string& name, 
             const std::string& date, 
             bool status,
-            uint8_t grade
+            uint16_t grade
         );
         virtual ~GradedTask() = default;
-        uint8_t getGrade() const;
-        void setGrade(uint8_t grade);
+        // Accessor & Mutator
+        uint16_t getGrade() const;
+        void setGrade(uint16_t grade);
 };
 
 #endif //GRADEDTASK_H
