+++ include/tasks/Homework.hpp
@@ -1,18 +1,23 @@
 #ifndef HOMEWORK_H
 #define HOMEWORK_H
 
+// Task header file
 #include "tasks/GradedTask.hpp"
+
+// Built in libraries
 #include <string>
 #include <cstdint>
 
+
 class Homework : public GradedTask {
     public:
+        // Constructor
         Homework(
             int idNum, 
             const std::string& name, 
             const std::string& date, 
             bool status,
-            uint8_t grade
+            uint16_t grade
         );
 };
