+++ include/tasks/Lab.hpp
@@ -1,18 +1,23 @@
 #ifndef LAB_H
 #define LAB_H
 
+// Task header file
 #include "tasks/GradedTask.hpp"
+
+// Built in libraries
 #include <string>
 #include <cstdint>
 
+
 class Lab : public GradedTask {
     public:
+        // Constructor
         Lab(
             int idNum, 
             const std::string& name, 
             const std::string& date, 
             bool status,
-            uint8_t grade
+            uint16_t grade
         );
 };
