+++ include/tasks/LabAssignment.hpp
@@ -1,18 +1,23 @@
 #ifndef LABASSIGNMENT_H
 #define LABASSIGNMENT_H
 
+// Task header file
 #include "tasks/GradedTask.hpp"
+
+// Built in libraries
 #include <string>
 #include <cstdint>
 
+
 class LabAssignment : public GradedTask {
     public:
+        // Constructor
         LabAssignment(
             int idNum, 
             const std::string& name, 
             const std::string& date, 
             bool status,
-            uint8_t grade
+            uint16_t grade
         );
 };
