+++ include/tasks/Project.hpp
@@ -1,18 +1,23 @@
 #ifndef PROJECT_H
 #define PROJECT_H
 
+// Task header file
 #include "tasks/GradedTask.hpp"
+
+// Built in library
 #include <string>
 #include <cstdint>
 
+
 class Project : public GradedTask {
     public:
+        // Constructor
         Project(
             int idNum, 
             const std::string& name, 
             const std::string& date, 
             bool status,
-            uint8_t grade
+            uint16_t grade
         );
 };
