+++ include/tasks/Reading.hpp
@@ -1,18 +1,23 @@
 #ifndef READING_H
 #define READING_H
 
+// Task header file
 #include "tasks/GradedTask.hpp"
+
+// Built in libraries
 #include <string>
 #include <cstdint>
 
+
 class Reading : public GradedTask {
     public:
+        // Constructor
         Reading(
             int idNum, 
             const std::string& name, 
             const std::string& date, 
             bool status,
-            uint8_t grade
+            uint16_t grade
         );
 };
