+++ src/tasks/GradedTask.cpp
@@ -1,17 +1,37 @@
+// Task header file
 #include "tasks/GradedTask.hpp"
 
+
+/**
+ * @brief A constructor for the class.
+ * @param idNum The identification number for the task.
+ * @param name The name of the task.
+ * @param date The date the task needs to be completed
+ * by in MM/DD/YYYY format.
+ * @param status Whether or not the task has been completed.
+ * @param grade The grade associated with the task
+ * if applicable.
+ */
 GradedTask::GradedTask(int idNum, const std::string& name, 
-    const std::string& date, bool status, uint8_t grade)
+    const std::string& date, bool status, uint16_t grade)
     : Task(idNum, name, date, status) {
 
     this->type = "GradedTask";
     this->grade = grade;
 }
 
-uint8_t GradedTask::getGrade() const {
+/**
+ * @brief An accessor for the grade field.
+ * @return The grade field as an integer.
+ */
+uint16_t GradedTask::getGrade() const {
     return grade;
 }
 
-void GradedTask::setGrade(uint8_t grade) {
+/**
+ * @brief A mutator for the grade field.
+ * @param grade The new grade integer to be set.
+ */
+void GradedTask::setGrade(uint16_t grade) {
     this->grade = grade;
 }
